Fix regression introduced with hash table fix
- This commit resolves a regression introduced whereby
undefined behavior was invoked with checking equality between
the keys in the `mutex_map` of `csystem`. Before this commit,
the `pthread_mutex_t*` was dereferenced as a `mutex_ref`.

+++ src/hashtable.h
@@ -9,7 +9,7 @@
 typedef uint64_t hash_t;
 typedef hash_t(*hash_function)(void*);
 typedef bool(*hash_equality_function)(void*, void*);
-#define REHASH_FACTOR (0.5)
+#define REHASH_FACTOR ((double)0.5)
 
 STRUCT_DECL(hash_table);
